+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -6,12 +6,17 @@
 
 - Add live metrics collection of requests/dependencies/exceptions
     ([#34673](https://github.com/Azure/azure-sdk-for-python/pull/34673))
+- Add live metrics collection of cpu time/process memory
+    ([#34735](https://github.com/Azure/azure-sdk-for-python/pull/34735))
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+- Updated FastAPI sample
+    ([#34738](https://github.com/Azure/azure-sdk-for-python/pull/34738))
 
 ## 1.0.0b23 (2024-02-28)